requirements.txt
Django
djangorestframework
psycopg2
ipython
drf-yasg
requests